ref: extract DropdownItemTextHighlighter component
