+++ frontend/src/components/Combobox/DropdownItem.tsx
@@ -1,17 +1,8 @@
 import { useMemo } from 'react'
-import Highlighter from 'react-highlight-words'
-import {
-  chakra,
-  CSSObject,
-  Flex,
-  Icon,
-  ListItem,
-  Stack,
-  Text,
-} from '@chakra-ui/react'
+import { CSSObject, Flex, Icon, ListItem, Stack, Text } from '@chakra-ui/react'
 import { UseComboboxPropGetters } from 'downshift'
-import escapeRegExp from 'lodash/escapeRegExp'
 
+import { DropdownItemTextHighlighter } from './DropdownItemTextHighlighter'
 import { ComboboxItem } from './types'
 import {
   isItemDisabled,
@@ -31,26 +22,6 @@ export interface DropdownItemProps {
   /** Current input value in dropdown for highlighting of matched text */
   inputValue: string
 }
-
-const DropdownHighlightMark = ({
-  showHoverBg,
-  children,
-  ...props
-}: {
-  showHoverBg: boolean
-  children: string
-}) => (
-  <chakra.mark
-    bg={showHoverBg ? 'primary.200' : 'primary.100'}
-    transitionProperty="background"
-    transitionDuration="ultra-fast"
-    transitionTimingFunction="ease-in"
-    color="primary.500"
-    {...props}
-  >
-    {children}
-  </chakra.mark>
-)
 
 export const DropdownItem = ({
   item,
@@ -72,15 +43,6 @@ export const DropdownItem = ({
     [item],
   )
 
-  /**
-   * Allows for fuzzy matching of searched characters, resulting in better UX.
-   * E.g. searching for `'rb'` will highlight `r` and `b` in `"radio button"`.
-   */
-  const regexSearchWords = useMemo(
-    () => [new RegExp(`[${escapeRegExp(inputValue)}]`, 'gi')],
-    [inputValue],
-  )
-
   return (
     <ListItem
       sx={itemStyles}
@@ -99,18 +61,11 @@ export const DropdownItem = ({
       <Flex flexDir="column">
         <Stack direction="row" align="center" spacing="1rem">
           {itemMeta.icon ? <Icon as={itemMeta.icon} sx={iconStyles} /> : null}
-          <Text>
-            <Highlighter
-              searchWords={regexSearchWords}
-              highlightTag={({ children }) => (
-                <DropdownHighlightMark
-                  children={children}
-                  showHoverBg={isSelected && !isActive}
-                />
-              )}
-              textToHighlight={itemMeta.label}
-            />
-          </Text>
+          <DropdownItemTextHighlighter
+            inputValue={inputValue}
+            showHoverBg={isSelected && !isActive}
+            textToHighlight={itemMeta.label}
+          />
         </Stack>
         <Text
           textStyle="body-2"
